Refactor transpiler repository (#1879)
## Changes

Following from #1861, this PR splits out `TranspilerRepository` into its
own module.

### Caveats/things to watch out for when reviewing

The scope of this PR is limited to moving the code.

### Linked issues

Follows: #1861

### Tests

- existing (and updated) integration tests
