+++ tests/integration/install/test_install_and_run.py
@@ -8,8 +8,9 @@ from email.parser import Parser as EmailParser
 import pytest
 
 from databricks.labs.lakebridge.config import TranspileConfig, TranspileResult
-from databricks.labs.lakebridge.install import TranspilerRepository, WheelInstaller, MavenInstaller
+from databricks.labs.lakebridge.install import WheelInstaller, MavenInstaller
 from databricks.labs.lakebridge.transpiler.lsp.lsp_engine import LSPEngine
+from databricks.labs.lakebridge.transpiler.repository import TranspilerRepository
 from databricks.labs.lakebridge.transpiler.transpile_engine import TranspileEngine
