+++ tests/integration/transpile/test_bladebridge.py
@@ -5,8 +5,9 @@ from pathlib import Path
 from databricks.sdk import WorkspaceClient
 
 from databricks.labs.lakebridge.config import TranspileConfig
-from databricks.labs.lakebridge.install import TranspilerRepository, WheelInstaller
+from databricks.labs.lakebridge.install import WheelInstaller
 from databricks.labs.lakebridge.transpiler.execute import transpile
+from databricks.labs.lakebridge.transpiler.repository import TranspilerRepository
 from databricks.labs.lakebridge.transpiler.transpile_engine import TranspileEngine
 
 logger = logging.getLogger(__name__)
